refactor(runtime/llm): extract tool-protocol helpers to own module
Moves the ~210 LOC tool-turn state machine helpers out of
chat-executor-tool-loop.ts into a new chat-executor-tool-protocol-helpers.ts.
chat-executor-tool-loop.ts drops from 2,550 -> 2,342 LOC. Net total
across the two files goes down slightly because clearing dead imports
from the main loop removed more than the extraction added.

Moved exports:
- syncToolProtocolSnapshot
- emitToolProtocolViolation
- pushToolResultMessage
- materializeResponseToolCalls
- sealPendingToolProtocol
- failClosedOnMalformedToolContinuation

TOOL_PROTOCOL_REPAIR_ERROR constant is now file-private in the new
module.

The new module imports ToolLoopCallbacks as `import type` from
chat-executor-tool-loop.ts — type-only so no runtime cycle. The main
loop file's now-unused imports (sanitizeToolCallsForReplay,
updateRuntimeContractToolProtocolSnapshot, noteToolProtocolRepair,
noteToolProtocolViolation, openToolProtocolTurn,
responseHasMalformedToolFinish, ToolProtocolRepairReason) are
dropped in the same commit.

Purely a file-organization change — zero behavior delta. Typecheck
green. llm/ test suite (720 tests, 1 pre-existing token-budget
compaction failure unchanged from main) matches baseline.
