+++ runtime/src/llm/chat-executor-tool-loop.ts
@@ -60,6 +60,14 @@ import {
   updateFailedToolStreak,
 } from "./chat-executor-failed-tool-tracking.js";
 import {
+  emitToolProtocolViolation,
+  failClosedOnMalformedToolContinuation,
+  materializeResponseToolCalls,
+  pushToolResultMessage,
+  sealPendingToolProtocol,
+  syncToolProtocolSnapshot,
+} from "./chat-executor-tool-protocol-helpers.js";
+import {
   applyActiveRoutedToolNames,
   buildActiveRoutedToolSet,
 } from "./chat-executor-routing-state.js";
@@ -81,7 +89,6 @@ import {
 } from "./chat-executor-continuation.js";
 import { evaluateShellWorkspaceWritePolicy } from "./shell-write-policy.js";
 import {
-  sanitizeToolCallsForReplay,
   generateFallbackContent,
   buildPromptToolContent,
 } from "./chat-executor-text.js";
@@ -133,18 +140,12 @@ import {
 import {
   type CompletionValidatorId,
   updateRuntimeContractValidatorSnapshot,
-  updateRuntimeContractToolProtocolSnapshot,
 } from "../runtime-contract/types.js";
 import {
   getPendingToolProtocolCalls,
   hasPendingToolProtocol,
-  noteToolProtocolRepair,
-  noteToolProtocolViolation,
-  openToolProtocolTurn,
   recordToolProtocolResult,
-  responseHasMalformedToolFinish,
   responseHasToolCalls,
-  type ToolProtocolRepairReason,
 } from "./tool-protocol-state.js";
 import {
   type RequestTaskObservationResult,
@@ -210,7 +211,6 @@ export interface ToolLoopCallbacks {
   serializeRemainingRequestMs(remainingRequestMs: number): number | null;
 }
 
-const TOOL_PROTOCOL_REPAIR_ERROR = "tool_protocol_repair";
 const TERMINAL_MUTATION_TOOL_NAMES = new Set([
   "system.applyPatch",
   "system.appendFile",
@@ -243,214 +243,6 @@ function buildToolLoopTerminalResult(
     runtimeContractSnapshot: ctx.runtimeContractSnapshot,
     mutationDetected: detectSuccessfulWorkspaceMutation(ctx.allToolCalls),
   };
-}
-
-function syncToolProtocolSnapshot(ctx: ExecutionContext): void {
-  ctx.runtimeContractSnapshot = updateRuntimeContractToolProtocolSnapshot({
-    snapshot: ctx.runtimeContractSnapshot,
-    open: hasPendingToolProtocol(ctx.toolProtocolState),
-    pendingToolCallIds: getPendingToolProtocolCalls(ctx.toolProtocolState).map(
-      (toolCall) => toolCall.id,
-    ),
-    repairCount: ctx.toolProtocolState.repairCount,
-    lastRepairReason: ctx.toolProtocolState.lastRepairReason,
-    violationCount: ctx.toolProtocolState.violationCount,
-    lastViolation: ctx.toolProtocolState.lastViolation,
-  });
-}
-
-function emitToolProtocolViolation(
-  ctx: ExecutionContext,
-  callbacks: ToolLoopCallbacks,
-  reason: string,
-  payload: Record<string, unknown> = {},
-): void {
-  noteToolProtocolViolation(ctx.toolProtocolState, reason);
-  syncToolProtocolSnapshot(ctx);
-  callbacks.emitExecutionTrace(ctx, {
-    type: "tool_protocol_violation",
-    phase: "tool_followup",
-    callIndex: ctx.callIndex,
-    payload: {
-      reason,
-      ...payload,
-    },
-  });
-}
-
-function pushToolResultMessage(params: {
-  readonly ctx: ExecutionContext;
-  readonly callbacks: ToolLoopCallbacks;
-  readonly toolCallId: string;
-  readonly toolName: string;
-  readonly content: string;
-  readonly args: Record<string, unknown>;
-  readonly isError: boolean;
-  readonly durationMs: number;
-  readonly synthetic?: boolean;
-  readonly protocolRepairReason?: ToolProtocolRepairReason;
-  readonly failureBudgetExempt?: boolean;
-}): void {
-  const {
-    ctx,
-    callbacks,
-    toolCallId,
-    toolName,
-    content,
-    args,
-    isError,
-    durationMs,
-    synthetic,
-    protocolRepairReason,
-  } = params;
-  callbacks.pushMessage(
-    ctx,
-    {
-      role: "tool",
-      content,
-      toolCallId,
-      toolName,
-    },
-    "tools",
-  );
-  callbacks.appendToolRecord(ctx, {
-    name: toolName,
-    args,
-    result: content,
-    isError,
-    durationMs,
-    toolCallId,
-    ...(synthetic ? { synthetic: true } : {}),
-    ...(protocolRepairReason ? { protocolRepairReason } : {}),
-    ...(params.failureBudgetExempt ? { failureBudgetExempt: true } : {}),
-  });
-  recordToolProtocolResult(ctx.toolProtocolState, toolCallId);
-  syncToolProtocolSnapshot(ctx);
-  callbacks.emitExecutionTrace(ctx, {
-    type: "tool_protocol_result_recorded",
-    phase: "tool_followup",
-    callIndex: ctx.callIndex,
-    payload: {
-      toolCallId,
-      tool: toolName,
-      synthetic: synthetic === true,
-      pendingToolCallIds: getPendingToolProtocolCalls(ctx.toolProtocolState).map(
-        (toolCall) => toolCall.id,
-      ),
-      ...(protocolRepairReason ? { protocolRepairReason } : {}),
-    },
-  });
-}
-
-function materializeResponseToolCalls(
-  ctx: ExecutionContext,
-  callbacks: ToolLoopCallbacks,
-): readonly LLMToolCall[] {
-  if (!ctx.response || !responseHasToolCalls(ctx.response)) {
-    return [];
-  }
-  if (hasPendingToolProtocol(ctx.toolProtocolState)) {
-    return ctx.response.toolCalls;
-  }
-
-  callbacks.pushMessage(
-    ctx,
-    {
-      role: "assistant",
-      content: ctx.response.content,
-      phase: "commentary",
-      toolCalls: sanitizeToolCallsForReplay(ctx.response.toolCalls),
-    },
-    "assistant_runtime",
-  );
-  openToolProtocolTurn(ctx.toolProtocolState, ctx.response.toolCalls);
-  syncToolProtocolSnapshot(ctx);
-  callbacks.emitExecutionTrace(ctx, {
-    type: "tool_protocol_opened",
-    phase: "tool_followup",
-    callIndex: ctx.callIndex,
-    payload: {
-      toolCallIds: ctx.response.toolCalls.map((toolCall) => toolCall.id),
-      toolNames: ctx.response.toolCalls.map((toolCall) => toolCall.name),
-      finishReason: ctx.response.finishReason,
-    },
-  });
-  return ctx.response.toolCalls;
-}
-
-function sealPendingToolProtocol(
-  ctx: ExecutionContext,
-  callbacks: ToolLoopCallbacks,
-  reason: ToolProtocolRepairReason,
-): boolean {
-  const pendingToolCalls = getPendingToolProtocolCalls(ctx.toolProtocolState);
-  if (pendingToolCalls.length === 0) {
-    return false;
-  }
-
-  for (const toolCall of pendingToolCalls) {
-    pushToolResultMessage({
-      ctx,
-      callbacks,
-      toolCallId: toolCall.id,
-      toolName: toolCall.name,
-      content: JSON.stringify({
-        error: "Runtime closed unresolved tool call before continuation",
-        code: TOOL_PROTOCOL_REPAIR_ERROR,
-        reason,
-      }),
-      args: {},
-      isError: true,
-      durationMs: 0,
-      synthetic: true,
-      protocolRepairReason: reason,
-      failureBudgetExempt: true,
-    });
-  }
-
-  noteToolProtocolRepair(ctx.toolProtocolState, reason);
-  if (ctx.response && responseHasToolCalls(ctx.response)) {
-    ctx.response = {
-      ...ctx.response,
-      content: "",
-    };
-  }
-  syncToolProtocolSnapshot(ctx);
-  callbacks.emitExecutionTrace(ctx, {
-    type: "tool_protocol_repaired",
-    phase: "tool_followup",
-    callIndex: ctx.callIndex,
-    payload: {
-      reason,
-      repairedToolCallIds: pendingToolCalls.map((toolCall) => toolCall.id),
-      repairedToolNames: pendingToolCalls.map((toolCall) => toolCall.name),
-    },
-  });
-  return true;
-}
-
-function failClosedOnMalformedToolContinuation(
-  ctx: ExecutionContext,
-  callbacks: ToolLoopCallbacks,
-): boolean {
-  if (!responseHasMalformedToolFinish(ctx.response)) {
-    return false;
-  }
-
-  const detail =
-    "Provider returned finishReason \"tool_calls\" without any tool calls; refusing to continue with an invalid tool-turn state.";
-  emitToolProtocolViolation(ctx, callbacks, "missing_tool_calls_for_finish_reason", {
-    finishReason: ctx.response?.finishReason,
-    contentPreview: (ctx.response?.content ?? "").slice(0, 240),
-  });
-  callbacks.setStopReason(ctx, "validation_error", detail);
-  if (ctx.response) {
-    ctx.response = {
-      ...ctx.response,
-      content: "",
-    };
-  }
-  return true;
 }
 
 function asDelegationOutputValidationCode(
